Done, with commented code

+++ src/components/course-editor/editable-item.js
@@ -1,5 +1,5 @@
 import React, {useState} from 'react'
-import {NavLink, Link, useLocation, useRouteMatch} from "react-router-dom";
+import {Link} from "react-router-dom";
 
 const EditableItem = (
     {
@@ -23,7 +23,7 @@ const EditableItem = (
         <>
             { !editing &&
                 <>
-                    <Link className={`btn text-left text-truncate d-inline-block col-12 ${active ? "active text-white" : ""}`}
+                    <Link className={`p-2 mt-2 text-left text-truncate d-inline-block col-12 ${active ? "active text-white" : ""}`}
                           to={path}>
                         {item.title}
                         <i onClick={(event) =>
@@ -31,35 +31,35 @@ const EditableItem = (
                             event.preventDefault()
                             setEditing(true)
                         }}
-                           className="fas fa-edit fa-lg float-right pt-1"/>
+                           className="fas text-center fa-edit fa-lg float-right"/>
                     </Link>
                 </>
             }
             { editing &&
-                <div className="col-12 bg-primary btn text-left">
-                <input className="form-control col-7 custom-control-inline"
-                   onChange={(event) => setCachedItem(
-                       {...cachedItem, title: event.target.value}
-                   )}
-                   value={cachedItem.title}>
-                </input>
-                <span className="float-right pt-2">
-                    <i onClick={() => {
-                        setEditing(false)
-                        updateItem(cachedItem)
-                    }}
-                       className={`fas fa-check fa-lg mr-3 text-white`}/>
-                    <i onClick={() => {
-                        // if (!active) {
+                <div className={`py-2 col-12 rounded ${active ? "" : "bg-dark"} text-left d-inline-flex`}>
+                    <input className="form-control col-8"
+                       onChange={(event) => setCachedItem(
+                           {...cachedItem, title: event.target.value}
+                       )}
+                       value={cachedItem.title}>
+                    </input>
+                    <span className="col-5 pt-2 d-inline-flex justify-content-end">
+                        <i onClick={() => {
                             setEditing(false)
-                            deleteItem(cachedItem)
-                        // }
-                        // else{
-                            alert(`Invalid delete: ${item.title}`)
-                        // }
-                    }}
-                       className={`fas fa-minus-circle fa-lg text-white`}/>
-                </span>
+                            updateItem(cachedItem)
+                        }}
+                           className={`fas fa-check fa-lg mr-3 text-white`}/>
+                        <i onClick={() => {
+                            if (!active) {
+                                setEditing(false)
+                                deleteItem(cachedItem)
+                            }
+                            else{
+                                alert(`Invalid delete: ${item.title}`)
+                            }
+                        }}
+                           className={`fas fa-minus-circle fa-lg text-white`}/>
+                    </span>
                 </div>
             }
         </>
